% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_lag_to_indep.R
\name{lor_lag_to_indep}
\alias{lor_lag_to_indep}
\title{Identify lag at which serial dependence is no longer present}
\usage{
lor_lag_to_indep(data, n_knots = 5, plot_results = TRUE, outDir = "",
  plot_title = "", plot_x_title = "Lag")
}
\arguments{
\item{data}{output of \code{\link{lorelogram}} function.}

\item{n_knots}{numeric. Number of knots in the smoothing cubic spline (see splines::bs for details).}

\item{plot_results}{logical. Create a .jpg plot of the results (default: TRUE)?}

\item{outDir}{character. Directory into which .csv and plot file are saved.}

\item{plot_title}{character. Title of the plot (default: NULL).}

\item{plot_x_title}{character. Title of x-axis of the plot (default: Lag).}
}
\value{
The function returns the minimum interval length necessary to approximate independence in the data.
}
\description{
This function identifies the spatial or temporal lag at which serial dependence is no longer present in binary data by determining when the first derivative of the pairwise log-odds ratios with respect to \eqn{\Delta}\emph{t} reaches zero. It allows for non-zero asymptotic log-odds ratios due to non-stationarity in the mean process or among-site variation.
}
\details{
\code{data} should be a data.frame containing the output of the \code{\link{lorelogram}} function.

First, a cubic spline is fitted to the pairwise log-odds ratios estimated by the \code{\link{lorelogram}} function at the series of spatial or temporal intervals of increasing length. Then, the first derivative of the spline curve is calculated. The lag at which the first derivative is as close to 0 (in absolute value) is returned. Users should visually inspect the plot of the first derivative function to ensure this lag results in a minimum.
}
\examples{
data(GrayFox_Hour)
lor <- lorelogram(GrayFox_Hour, max_lag = 30)
lor_lag_to_indep(lor)


}
