% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_plot.R
\name{lor_plot}
\alias{lor_plot}
\title{Plot pairwise log-odds ratios at intervals of increasing length}
\usage{
lor_plot(data, save_LOR_plot = FALSE, outDir = "", plot_title = "",
  plot_x_title = "Lag")
}
\arguments{
\item{data}{output of \code{\link{lorelogram}} function.}

\item{save_LOR_plot}{logical. Create a .jpg plot of the results (default: FALSE)?}

\item{outDir}{character. Directory into which .csv and plot file are saved.}

\item{plot_title}{character. Title of the plot (default: NULL).}

\item{plot_x_title}{character. Title of x-axis of the plot (default: Lag).}
}
\value{
The fucntion returns a plot of the estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag}.
}
\description{
The estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag} calculated by the \code{\link{lorelogram}} function are plotted against (spatial or temporal) intervals of increasing length. The plot of the estimates versus lags provides a graphical description of how correlation between outcomes x-lag apart changes at the increase of the distance (in space or time) between the sampling replicates and it allows to identify which spatial or temporal interval is necessary to approximate independence in binary data.
}
\details{
\code{data} should resemble a binary detection/nondetection history matrix such that provided as an output by the function \code{\link[camtrapR:detectionHistory]{camtrapR::detectionHistory}}. \code{\link{lorelogram}} can handle NAs in \code{data}.
}
\examples{
data(GrayFox_Hour)
lor <- lorelogram(GrayFox_Hour, max_lag = 120)
lor_plot(lor)

}
