% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorelogram.R
\name{lorelogram}
\alias{lorelogram}
\title{Calculate and plot pairwise log-odds ratios at intervals of increasing length}
\usage{
lorelogram(data, max_lag = 30, write_csv = TRUE, outDir = "",
  plot_LOR = TRUE, plot_title = "", plot_x_title = "Lag")
}
\arguments{
\item{data}{data.frame containing binary data in the wide format, with rows representing sampling units (e.g., camera trap sites or transects) and columns representing repeated samplings (e.g., temporal occasions or spatial replicates).}

\item{max_lag}{numeric. The maximum spatial or temporal lag between two sampling occasions at the same sampling units (default: 30).}

\item{write_csv}{logical. Should output be saved as a .csv (default: TRUE)?}

\item{outDir}{character. Directory into which .csv and plot file are saved.}

\item{plot_LOR}{logical. Create a .jpg plot of the results (default: TRUE)?}

\item{plot_title}{character. Title of the plot (default: NULL)}

\item{plot_x_title}{character. Title of x-axis of the plot (default: Lag)}
}
\value{
A data.frame containing estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag} is returned.
}
\description{
This function estimates pairwise log-odds ratios in binary data at intervals of increasing length between subsequent sampling replicates. Plot of the estimates versus lags provides a graphical description of how correlation between outcomes x-lag apart changes at the increase of the distance (in space or time) between the sampling replicates.
}
\details{
\code{data} should resemble a binary detection/nondetection history matrix such that provided as an output by the function \code{\link[camtrapR:detectionHistory]{camtrapR::detectionHistory}}. \code{\link{lorelogram}} can handle NAs in \code{data}.
}
\examples{
data(GrayFox_Hour)
lorelogram(GrayFox_Hour)

}
