% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_lag_to_indep.R
\name{lor_lag_to_indep}
\alias{lor_lag_to_indep}
\title{Identify lag to approximate independence in binary data}
\usage{
lor_lag_to_indep(data, n_knots = 5, plot_results = TRUE, outDir = "",
  plot_title = "", plot_x_title = "Lag")
}
\arguments{
\item{data}{output of \code{\link{lorelogram}} function.}

\item{n_knots}{numeric. Number of knots in the smoothing cubic spline (see splines::bs for details).}

\item{plot_results}{logical. Create a .jpg plot of the results (default: TRUE)?}

\item{outDir}{character. Directory into which .csv and plot file are saved.}

\item{plot_title}{character. Title of the plot (default: NULL).}

\item{plot_x_title}{character. Title of x-axis of the plot (default: Lag).}
}
\value{
The function returns the minimum interval length necessary to approximate independence in the data.
}
\description{
This function identifies spatial or temporal intervals required to approximate independence in binary data by determining when the first derivative of the log-odds ratios with respect to \eqn{\Delta}\emph{t} reaches zero. It allows for non-zero asymptotic values due to among-site variation or cyclic patterns.
}
\details{
\code{data} should be a data.frame containing the output of the \code{\link{lorelogram}} function.

First, a smoothing cubic spline is fitted to the log-odds ratios values estimated by the \code{\link{lorelogram}} function at the series of spatial or temporal intervals of increasing length. Then, the first derivative of the spline curve is calculated.

We recommend a visual inspection of the plot produced. When log-odds ratios present cyclic patterns (e.g., due to daily activity patterns affecting camera trap data) at the spatial or temporal scale used in the analysis, the \code{\link{lor_lag_to_indep}} can not be used to identify the minimum interval length that approximate independence in the data.
}
\examples{
data(GrayFox_Hour)
# cyclic pattern
lor <- lorelogram(GrayFox_Hour, max_lag = 120)
lor_lag_to_indep(lor)

# non-cyclic pattern
lor <- lorelogram(GrayFox_Hour, max_lag = 30)
lor_lag_to_indep(lor)


}
