% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_plot.R
\name{lor_plot}
\alias{lor_plot}
\title{Plot pairwise log-odds ratios at intervals of increasing length}
\usage{
lor_plot(data, save_LOR_plot = FALSE, outDir = "",
  colour = "#0C71C9", linetype = "solid", title = "",
  x_axis_title = "Lag", y_axis_title = "Log-odds ratio", ylim = NULL,
  x_break = 10, alpha = 1)
}
\arguments{
\item{data}{output of \code{\link{lorelogram}} function.}

\item{save_LOR_plot}{logical. Create a .jpg plot of the results (default: FALSE)?}

\item{outDir}{character. Directory into which .jpg plot is saved.}

\item{colour}{character. Color to fill the confidence interval band (defaul: "#0C71C9")}

\item{linetype}{character. Linetype for the average log-odds ratio values (default: "solid"). It accepts all the linetype available for ggplot2::geom_line.}

\item{title}{character. Title of the plot (default: NULL).}

\item{x_axis_title}{character. Title of x-axis of the plot (default: "Lag").}

\item{y_axis_title}{character. Title of y-axis of the plot (default: "Log-odds ratio").}

\item{ylim}{numeric. Vector of two values \code{c(y_min, y_max)} defining range for the y axis (default: NULL).}

\item{x_break}{numeric. Unit-lag distance between primary breaks in the x_axis (default: 10).}

\item{alpha}{numeric. Set transparency of the band describing the confidence interval. Should be a value between 0 and 1 (default: 1).}
}
\value{
The function returns a plot of the estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag}.
}
\description{
The estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag} calculated by the \code{\link{lorelogram}} function are plotted against (spatial or temporal) intervals of increasing length. The plot of the estimates versus lags provides a graphical description of how correlation between outcomes x-lag apart changes at the increase of the distance (in space or time) between the sampling replicates and it allows to identify which spatial or temporal interval is necessary to approximate independence in binary data.
}
\details{
\code{data} should resemble a binary detection/nondetection history matrix such that provided as an output by the function \code{\link[camtrapR:detectionHistory]{camtrapR::detectionHistory}}. \code{\link{lorelogram}} can handle NAs in \code{data}.
}
\examples{
data(GrayFox_Hour)
lor <- lorelogram(GrayFox_Hour, max_lag = 120)
lor_plot(lor)

}
