% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorelogram.R
\name{lorelogram}
\alias{lorelogram}
\title{Calculate pairwise log-odds ratios at intervals of increasing length}
\usage{
lorelogram(data, data_format = "wide", max_lag = 30,
  lor_type = "empirical", id_rand_eff = FALSE, bin_width = 1,
  plot_LOR = TRUE, write_csv = FALSE, outDir = "")
}
\arguments{
\item{data}{data.frame containing binary data in the wide format, with rows representing sampling units (e.g., camera trap sites or transects) and columns representing repeated samplings (e.g., temporal occasions or spatial replicates). See Details section below.}

\item{data_format}{character. Data can be provided in "wide" (default) or "long" format. See details for more information.}

\item{max_lag}{numeric. The maximum spatial or temporal lag between two sampling occasions at the same sampling units (default: 30) that should be considered.}

\item{lor_type}{character. Lorelogram can either be estimate using an "empirical" (default) or "model-based" approach.}

\item{id_rand_eff}{logical. Conditional on lor_type = "model-based". Does the model has to include sampling unit ID as random effect?}

\item{bin_width}{numeric. Number of lag that should be included in each bin. The default (=1) represents no binning.}

\item{plot_LOR}{logical. Create a plot of the results (default: TRUE)?}

\item{write_csv}{logical. Should the output be saved as a .csv (default: FALSE)?}

\item{outDir}{character. Directory into which .csv and plot file are saved.}
}
\value{
A data.frame containing estimates of pairwise log-odds ratios and associated 95\% confidence intervals for each lag between 1 and \code{max_lag} is returned.
}
\description{
This function estimates pairwise log-odds ratios in binary data at intervals of increasing length between subsequent sampling replicates. Plotting of the estimates versus lags (see the \code{\link{lor_plot}} function) provides a graphical description of how correlation between outcomes x-lag apart changes at the increase of the distance (in space or time) between the sampling replicates.
}
\details{
\code{\link{lorelogram}} can handle NAs in \code{data}. \code{data} should resemble a binary detection/nondetection history matrix such that provided as an output by the function \code{\link[camtrapR:detectionHistory]{camtrapR::detectionHistory}}. The first column should contain unitID (a unique identifier for each sampling units); each column from the second to the last should contain the binary data and should follow the spatial or temporal order in which the data were collected (e.g., second, third, and fourth columns should contain data from the first, second, third sampling replicates and so on). If unequal intervals are present in the data, fill gaps with columns of all NAs. Spatio- or temporal- difference between two subsequent columns (e.g. second and third) corresponds to 1-unit lag.
}
\examples{
data(GrayFox_Hour)
lorelogram(GrayFox_Hour, max_lag = 60)



}
\seealso{
\code{\href[CompRandFld]{EVariogram}},
}
